refactor: Break down monolithic cli.py into modular structure
  - Split 1,465-line cli.py into focused modules under
  ingenious/cli/
  - Created separate modules for logical command groups:
    * server_commands.py - Server operations (serve, prompt-tuner)
    * project_commands.py - Project management (init)
    * test_commands.py - Test execution
    * workflow_commands.py - Workflow management
    * help_commands.py - Help and status commands
    * utilities.py - CliFunctions and utility functions
  - Maintained full backward compatibility with legacy cli.py
  import
  - Added __main__.py for direct module execution support
  - Fixed markpickle → jsonpickle import error in model_utils.py

  Benefits:
  - Improved maintainability with single-responsibility modules
  - Enhanced readability (76-443 lines vs 1,465 lines)
  - Easier navigation and future extensions
  - All 201 tests passing, pre-commit hooks clean
