+++ ingenious/cli/__init__.py
@@ -1,0 +1,10 @@
+"""
+CLI package for Insight Ingenious.
+
+This package contains the command-line interface for the Insight Ingenious framework,
+organized into logical modules for better maintainability.
+"""
+
+from .main import app
+
+__all__ = ["app"]
