+++ ingenious/cli/main.py
@@ -1,0 +1,73 @@
+"""
+Main CLI application setup for Insight Ingenious.
+
+This module contains the core CLI application configuration and imports
+all command modules to register them with the typer app.
+"""
+
+from __future__ import annotations
+
+import typer
+from rich.console import Console
+from rich.theme import Theme
+
+from ingenious.utils.lazy_group import LazyGroup
+
+# ────────────────────────────────────────────────────────────────────────────
+# Typer application singleton
+# ────────────────────────────────────────────────────────────────────────────
+
+app: typer.Typer = typer.Typer(
+    cls=LazyGroup,
+    no_args_is_help=True,  # `ingen` → show help when no args
+    pretty_exceptions_show_locals=False,  # cleaner tracebacks in production
+    help="""
+🚀 Insight Ingenious - GenAI Accelerator
+
+A powerful framework for building and deploying AI agent workflows.
+
+Quick Start:
+  ingen init                    # Initialize a new project
+  ingen serve                   # Start the API server
+  ingen workflows               # List available workflows
+
+Common Commands:
+  init, serve, test, workflows, prompt-tuner
+
+Data Processing:
+  dataprep, document-processing
+
+Get help for any command with: ingen <command> --help
+    """.strip(),
+)
+
+custom_theme = Theme(
+    {
+        "info": "dim cyan",
+        "warning": "dark_orange",
+        "danger": "bold red",
+        "error": "bold red",
+        "debug": "khaki1",
+    }
+)
+
+console = Console(theme=custom_theme)
+
+# Import command modules to register them with the app
+from . import (
+    help_commands,
+    project_commands,
+    server_commands,
+    test_commands,
+    workflow_commands,
+)
+
+# Register commands from each module
+server_commands.register_commands(app, console)
+project_commands.register_commands(app, console)
+test_commands.register_commands(app, console)
+workflow_commands.register_commands(app, console)
+help_commands.register_commands(app, console)
+
+if __name__ == "__main__":
+    app()
