+++ ingenious/cli/server_commands.py
@@ -1,0 +1,288 @@
+"""
+Server-related CLI commands for Insight Ingenious.
+
+This module contains commands for starting and managing the API server.
+"""
+
+from __future__ import annotations
+
+import importlib
+import os
+import pkgutil
+from pathlib import Path
+from sysconfig import get_paths
+from typing import Optional
+
+import typer
+import uvicorn
+from dotenv import load_dotenv
+from rich.console import Console
+from typing_extensions import Annotated
+
+from ingenious.cli.utilities import CliFunctions
+
+# Load environment variables from .env file
+load_dotenv()
+
+
+def register_commands(app: typer.Typer, console: Console) -> None:
+    """Register server-related commands with the typer app."""
+    
+    @app.command(name="serve", help="Start the API server with web interface")
+    def serve(
+        config: Annotated[
+            Optional[str],
+            typer.Option(
+                "--config",
+                "-c",
+                help="Path to config.yml file (default: ./config.yml or $INGENIOUS_PROJECT_PATH)",
+            ),
+        ] = None,
+        profile: Annotated[
+            Optional[str],
+            typer.Option(
+                "--profile",
+                "-p",
+                help="Path to profiles.yml file (default: ./profiles.yml or $INGENIOUS_PROFILE_PATH)",
+            ),
+        ] = None,
+        host: Annotated[
+            str,
+            typer.Option("--host", "-h", help="Host to bind the server (default: 0.0.0.0)"),
+        ] = "0.0.0.0",
+        port: Annotated[
+            int,
+            typer.Option(
+                "--port", help="Port to bind the server (default: 80 or $WEB_PORT)"
+            ),
+        ] = int(os.getenv("WEB_PORT", "80")),
+        no_prompt_tuner: Annotated[
+            bool,
+            typer.Option("--no-prompt-tuner", help="Disable the prompt tuner interface"),
+        ] = False,
+    ):
+        """
+        🚀 Start the Insight Ingenious API server with web interface.
+
+        The server provides:
+        • REST API endpoints for agent workflows
+        • Chainlit chat interface at /chainlit
+        • Prompt tuning interface at /prompt-tuner (unless disabled)
+
+        AVAILABLE WORKFLOWS & CONFIGURATION REQUIREMENTS:
+
+        ✅ Minimal Configuration (Azure OpenAI only):
+          • classification-agent - Route input to specialized agents
+          • bike-insights - Sample domain-specific workflow
+
+        🔍 Requires Azure Search Services:
+          • knowledge-base-agent - Search knowledge bases
+
+        📊 Requires Database Configuration:
+          • sql-manipulation-agent - Execute SQL queries
+
+        📄 Optional Azure Document Intelligence:
+          • document-processing - Extract text from PDFs/images
+
+        QUICK TEST:
+          curl -X POST http://localhost:{port}/api/v1/chat \\
+            -H "Content-Type: application/json" \\
+            -d '{{"user_prompt": "Hello", "conversation_flow": "classification-agent"}}'
+
+        For detailed configuration: ingen workflows --help
+        """
+        return run_rest_api_server(
+            project_dir=config,
+            profile_dir=profile,
+            host=host,
+            port=port,
+            run_dir="",
+            enable_prompt_tuner=not no_prompt_tuner,
+        )
+
+    # Keep old command for backward compatibility
+    @app.command(hidden=True)
+    def run_rest_api_server(
+        project_dir: Annotated[
+            str,
+            typer.Argument(help="The path to the config file. "),
+        ] = None,
+        profile_dir: Annotated[
+            str,
+            typer.Argument(
+                help="The path to the profile file. If left blank it will use './profiles.yml' if it exists, otherwise '$HOME/.ingenious/profiles.yml'"
+            ),
+        ] = None,
+        host: Annotated[
+            str,
+            typer.Argument(
+                help="The host to run the server on. Default is 0.0.0.0. For local development outside of docker use 127.0.0.1"
+            ),
+        ] = "0.0.0.0",
+        port: Annotated[
+            int,
+            typer.Argument(help="The port to run the server on. Default is 80."),
+        ] = 80,
+        run_dir: Annotated[
+            str,
+            typer.Argument(help="The directory in which to launch the web server."),
+        ] = "",
+        enable_prompt_tuner: Annotated[
+            bool,
+            typer.Option(help="Enable the prompt tuner interface. Default is True."),
+        ] = True,
+    ):
+        """
+        Run a FastAPI server that presents your agent workflows via REST endpoints.
+
+        AVAILABLE WORKFLOWS & CONFIGURATION REQUIREMENTS:
+
+        ⭐ "Hello World" Workflow (Azure OpenAI only):
+          • bike-insights - **RECOMMENDED STARTING POINT** - Multi-agent bike sales analysis
+
+        ✅ Simple Text Processing (Azure OpenAI only):
+          • classification_agent - Route input to specialized agents
+
+        🔍 Requires Azure Search Services:
+          • knowledge_base_agent - Search knowledge bases
+
+        📊 Requires Database Configuration:
+          • sql_manipulation_agent - Execute SQL queries
+
+        📄 Optional Azure Document Intelligence:
+          • document-processing - Extract text from PDFs/images
+
+        For detailed configuration requirements, see:
+        docs/workflows/README.md
+
+        QUICK TEST (Hello World):
+        curl -X POST http://localhost:PORT/api/v1/chat \\
+          -H "Content-Type: application/json" \\
+          -d '{
+            "user_prompt": "{\\"stores\\": [{\\"name\\": \\"Hello Store\\", \\"location\\": \\"NSW\\", \\"bike_sales\\": [{\\"product_code\\": \\"HELLO-001\\", \\"quantity_sold\\": 1, \\"sale_date\\": \\"2023-04-01\\", \\"year\\": 2023, \\"month\\": \\"April\\", \\"customer_review\\": {\\"rating\\": 5.0, \\"comment\\": \\"Great first experience!\\"}}], \\"bike_stock\\": []}], \\"revision_id\\": \\"hello-1\\", \\"identifier\\": \\"world\\"}",
+            "conversation_flow": "bike-insights"
+          }'
+        """
+        if project_dir is not None:
+            os.environ["INGENIOUS_PROJECT_PATH"] = project_dir
+        elif os.getenv("INGENIOUS_PROJECT_PATH") is None:
+            # Default to config.yml in current directory
+            default_config_path = Path.cwd() / "config.yml"
+            if default_config_path.exists():
+                os.environ["INGENIOUS_PROJECT_PATH"] = str(default_config_path)
+                print(f"Using default config path: {default_config_path}")
+
+        if profile_dir is None:
+            # Check if environment variable is set
+            if os.getenv("INGENIOUS_PROFILE_PATH") is None:
+                # Try project directory first
+                project_profile_path = Path.cwd() / "profiles.yml"
+                if project_profile_path.exists():
+                    profile_dir = project_profile_path
+                    print(f"Using profiles.yml from project directory: {profile_dir}")
+                else:
+                    # Fall back to home directory
+                    home_dir = os.path.expanduser("~")
+                    profile_dir = Path(home_dir) / Path(".ingenious") / Path("profiles.yml")
+                    print(f"Using profiles.yml from home directory: {profile_dir}")
+            else:
+                profile_dir = Path(os.getenv("INGENIOUS_PROFILE_PATH"))
+        else:
+            profile_dir = Path(profile_dir)
+
+        print(f"Profile path: {profile_dir}")
+        os.environ["INGENIOUS_PROFILE_PATH"] = str(profile_dir).replace("\\", "/")
+        import ingenious.config.config as ingen_config
+
+        config = ingen_config.get_config()
+
+        # Override prompt tuner setting based on CLI flag
+        config.prompt_tuner.enable = enable_prompt_tuner
+
+        # Override host and port from CLI parameters
+        config.web_configuration.ip_address = host
+        config.web_configuration.port = port
+
+        # We need to clean this up and probably separate overall system config from fast api, eg. set the config here in cli and then pass it to FastAgentAPI
+        # As soon as we import FastAgentAPI, config will be loaded hence to ensure that the environment variables above are loaded first we need to import FastAgentAPI after setting the environment variables
+        from ingenious.main import FastAgentAPI
+
+        os.environ["LOADENV"] = "False"
+        console.print(f"Running all elements of the project in {project_dir}", style="info")
+        # If the code has been pip installed then recursively copy the ingenious folder into the site-packages directory
+        if CliFunctions.PureLibIncludeDirExists():
+            src = Path(os.getcwd()) / Path("ingenious/")
+            if os.path.exists(src):
+                CliFunctions.copy_ingenious_folder(
+                    src, Path(get_paths()["purelib"]) / Path("ingenious/")
+                )
+
+        print(f"Current working directory: {os.getcwd()}")
+
+        def print_namespace_modules(namespace):
+            package = importlib.import_module(namespace)
+            if hasattr(package, "__path__"):
+                for module_info in pkgutil.iter_modules(package.__path__):
+                    print(f"Found module: {module_info.name}")
+            else:
+                print(f"{namespace} is not a package")
+
+        os.environ["INGENIOUS_WORKING_DIR"] = str(Path(os.getcwd()))
+        os.chdir(str(Path(os.getcwd())))
+        print_namespace_modules(
+            "ingenious.services.chat_services.multi_agent.conversation_flows"
+        )
+
+        fast_agent_api = FastAgentAPI(config)
+
+        # Access the FastAPI app instance
+        app = fast_agent_api.app
+
+        # change directory to project dir
+        uvicorn.run(
+            app,
+            host=config.web_configuration.ip_address,
+            port=config.web_configuration.port,
+        )
+
+    @app.command(name="prompt-tuner", help="Start standalone prompt tuning interface")
+    def prompt_tuner(
+        port: Annotated[
+            int,
+            typer.Option("--port", "-p", help="Port for the prompt tuner (default: 5000)"),
+        ] = 5000,
+        host: Annotated[
+            str,
+            typer.Option(
+                "--host", "-h", help="Host to bind the prompt tuner (default: 127.0.0.1)"
+            ),
+        ] = "127.0.0.1",
+    ):
+        """
+        🎯 Start the standalone prompt tuning web interface.
+
+        The prompt tuner allows you to:
+        • Edit and test agent prompts
+        • Run batch tests with sample data
+        • Compare different prompt versions
+        • Download test results
+
+        Access the interface at: http://{host}:{port}
+
+        Note: This starts only the prompt tuner, not the full API server.
+        For the complete server with all interfaces, use: ingen serve
+        """
+        console.print(f"🎯 Starting prompt tuner at http://{host}:{port}")
+        console.print(
+            "💡 Tip: Use 'ingen serve' to start the full server with all interfaces"
+        )
+
+        # Import and start the Flask app for prompt tuning
+        try:
+            from ingenious_prompt_tuner.run_flask_app import app as flask_app
+
+            flask_app.run(host=host, port=port, debug=True)
+        except ImportError:
+            console.print("[red]❌ Prompt tuner dependencies not available[/red]")
+            console.print("Install with: uv add flask")
+            raise typer.Exit(1)
